Fixed Kernel config file location

+++ extra/shifter_cle6_kmod_deps.spec
@@ -7,7 +7,7 @@
 
 Name:           shifter_cle6_kmod_deps-%(uname -r)
 Version:        1.0
-Release:        1
+Release:        3
 License:        GPL
 BuildRequires: kernel-source kernel-syms
 BuildRoot:      %{_tmppath}/%{name}-%{version}-build
@@ -20,7 +20,15 @@ xfs, ext4 and deps
 
 %build
 cd /usr/src/linux
-cp arch/x86/configs/cray_ari_c_defconfig .config
+if [ -e "arch/x86/configs/cray_ari_c_defconfig" ]; then
+  # assuming we are on a chroot environment
+  cp arch/x86/configs/cray_ari_c_defconfig .config
+else
+  # assuming we are on a compute node
+  cp /proc/config.gz ./
+  gunzip config.gz
+  mv config .config
+fi
 make oldconfig
 make modules_prepare
 make M=fs/xfs CONFIG_XFS_FS=m
@@ -45,4 +53,8 @@ depmod -a
 %defattr(-,root,root)
 /lib/*
 
-
+%changelog
+* Mon Aug 08 2016 Miguel Gila <miguel.gila@cscs.ch> 1.0-3
+ - Added if case to support building on chroot environment
+* Wed Jul 20 2016 Miguel Gila <miguel.gila@cscs.ch> 1.0-2
+ - Fixed Kernel config file location
